Scripts update

+++ Project.Trunk/Weather.Station/scripts/weather.menu.sh
@@ -24,8 +24,11 @@ do
   read a
   if nocase "$a" "N"
   then
-    cd node
-    rm node.log
+    cd ../node
+    if [ -f node.log ]
+    then
+      rm node.log
+    fi
     nohup node weather.server.js > node.log &
     cd ..
     echo -n "Log is node/node.log. Hit [return] "
@@ -33,11 +36,15 @@ do
   elif nocase "$a" "W"
   then
     echo Make sure you have started the WebSocket server \(Option N\).
-    rm weather.station.log
+    if [ -f weather.station.log ]
+    then
+      rm weather.station.log
+    fi
     # ./weather.station.reader > weather.station.log
     nohup ./weather.station.reader.sh > weather.station.log &
     echo .
-		ADDR=`ifconfig wlan0 2> /dev/null  | awk '/inet addr:/ {print $2}' | sed 's/addr://'`
+		# ADDR=`ifconfig wlan0 2> /dev/null  | awk '/inet addr:/ {print $2}' | sed 's/addr://'`
+		ADDR=`hostname -I`
 		echo then from your browser, reach http://$ADDR:9876/data/weather.station/analog.all.html
 		echo IP is $(hostname -I)
     echo -n "Log is in weather.station.log. Hit [return] "
